processor queue

+++ kratos-processor/kratos-processor-api/src/main/java/com/andyadc/kratos/processor/api/concurrent/queue/mpmc/queue/ConcurrentQueue.java
@@ -13,9 +13,9 @@ public interface ConcurrentQueue<E> {
 
     E peek();
 
-    long size();
+    int size();
 
-    long capacity();
+    int capacity();
 
     boolean isEmpty();
